+++ code/packages/rust/html-lexer/src/generated_html1.rs
@@ -9428,7 +9428,8 @@ pub fn html1_lexer_definition() -> StateMachineDefinition {
             stack_pop: None,
             stack_push: Vec::new(),
             actions: vec![
-                "parse_error(abrupt-closing-of-empty-comment)".to_string(),
+                "parse_error(incorrectly-opened-comment)".to_string(),
+                "append_comment(-)".to_string(),
                 "emit_current_token".to_string(),
             ],
             consume: true,
@@ -9444,7 +9445,8 @@ pub fn html1_lexer_definition() -> StateMachineDefinition {
             stack_pop: None,
             stack_push: Vec::new(),
             actions: vec![
-                "parse_error(eof-in-comment)".to_string(),
+                "parse_error(incorrectly-opened-comment)".to_string(),
+                "append_comment(-)".to_string(),
                 "emit_current_token".to_string(),
                 "emit(EOF)".to_string(),
             ],
@@ -9461,6 +9463,7 @@ pub fn html1_lexer_definition() -> StateMachineDefinition {
             stack_pop: None,
             stack_push: Vec::new(),
             actions: vec![
+                "parse_error(incorrectly-opened-comment)".to_string(),
                 "append_comment(-)".to_string(),
                 "append_comment(current)".to_string(),
             ],